third_party/WebKit/LayoutTests/virtual/htxg-origin-trial-with-network-service/http/tests/loading/htxg/htxg-location-expected.txt

main frame - didStartProvisionalLoadForFrame
main frame - didCommitLoadForFrame
main frame - didReceiveTitle: Location of SignedHTTPExchange
main frame - didFinishDocumentLoadForFrame
main frame - didHandleOnloadEventsForFrame
main frame - didFinishLoadForFrame
frame "htxg_iframe" - didReceiveTitle: 
frame "htxg_iframe" - didStartProvisionalLoadForFrame
frame "htxg_iframe" - didFailProvisionalLoadWithError
This is a testharness.js-based test.
FAIL Location of SignedHTTPExchange promise_test: Unhandled rejection with value: "timeout"
Harness: the test ran to completion.
